+++ samples/kaspresso-sample/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
         </activity>
 
         <activity
-            android:name=".autoscrollfallback.AutoscrollFallbackActivity"
+            android:name=".scrollresolver.ScrollResolverActivity"
             android:label="@string/simple_fragment_title"/>
 
         <activity
