+++ samples/kaspresso-sample/src/main/res/layout/activity_main.xml
@@ -22,7 +22,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_below="@+id/activity_main_title"
-        android:text="@string/autoscroll_fallback_title" />
+        android:text="@string/scroll_resolver_title" />
 
     <Button
         android:id="@+id/activity_main_simple_sample_button"
