+++ samples/kaspresso-sample/src/main/res/values-ru/strings.xml
@@ -14,7 +14,7 @@
     <string name="main_screen_shared_test">Shared Test with Robolectric пример</string>
 
     <!-- Auto scroll fallback screen -->
-    <string name="autoscroll_fallback_title">Auto scroll fallback</string>
+    <string name="scroll_resolver_title">Auto scroll fallback</string>
 
     <!-- Simple fragment screen -->
     <string name="simple_fragment_title">Простой фрагмент</string>
